+++ module-gui/test/mock/TestListViewProvider.cpp
@@ -1,0 +1,63 @@
+#include "TestListViewProvider.hpp"
+#include <log/log.hpp>
+
+namespace gui
+{
+
+    TestListItem::TestListItem()
+    {
+        setMinimumSize(testStyle::item_w, testStyle::item_h);
+        setMaximumSize(testStyle::item_w, testStyle::item_h);
+    };
+
+    int TestListViewProvider::getItemCount() const
+    {
+        return testItemCount;
+    }
+
+    unsigned int TestListViewProvider::getMinimalItemHeight()
+    {
+
+        return testItemMinimalHeight;
+    }
+
+    void TestListViewProvider::requestRecords(const uint32_t offset, const uint32_t limit)
+    {
+        modelIndex     = 0;
+        internalOffset = offset;
+        internalLimit  = limit;
+        list->onProviderDataUpdate();
+    }
+
+    gui::ListItem *TestListViewProvider::getItem(gui::Order order)
+    {
+        unsigned int index = 0;
+        if (order == gui::Order::Previous) {
+            index = internalLimit - modelIndex - 1;
+        }
+        if (order == gui::Order::Next) {
+            index = internalOffset + modelIndex;
+        }
+
+        if (index < testItemCount) {
+
+            if (order == gui::Order::Previous && index < internalOffset) {
+                return nullptr;
+            }
+
+            auto testItem = new TestListItem();
+            testItem->ID  = index;
+
+            if (notEqualItems) {
+                testItem->setMinimumSize(testStyle::item_w, testStyle::item_h + testStyle::item_h * index);
+            }
+
+            modelIndex++;
+
+            return testItem;
+        }
+
+        return nullptr;
+    }
+
+}; // namespace gui